backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
langchain
langchain-google-genai
langsmith
langgraph
google-generativeai
httpx
sqlalchemy
pandas
openpyxl
psycopg2-binary
